+++ src/main/java/com/sprint/mission/discodeit/exception/readStatus/ReadStatusNotFoundException.java
@@ -1,10 +1,17 @@
-package com.sprint.mission.discodeit.exception.readStatus;
+package com.sprint.mission.discodeit.exception.readstatus;
 
 import com.sprint.mission.discodeit.exception.ErrorCode;
 
-public class ReadStatusNotFoundException extends ReadStatusException {
+import java.util.UUID;
 
-    public ReadStatusNotFoundException(ErrorCode errorCode) {
-        super(errorCode);
+public class ReadStatusNotFoundException extends ReadStatusException {
+    public ReadStatusNotFoundException() {
+        super(ErrorCode.READ_STATUS_NOT_FOUND);
+    }
+
+    public static ReadStatusNotFoundException withId(UUID readStatusId) {
+        ReadStatusNotFoundException exception = new ReadStatusNotFoundException();
+        exception.addDetail("readStatusId", readStatusId);
+        return exception;
     }
 }
