--- src/main/java/com/sprint/mission/discodeit/exception/readstatus/ReadStatusException.java
@@ -1,14 +1,0 @@
-package com.sprint.mission.discodeit.exception.readstatus;
-
-import com.sprint.mission.discodeit.exception.DiscodeitException;
-import com.sprint.mission.discodeit.exception.ErrorCode;
-
-public class ReadStatusException extends DiscodeitException {
-    public ReadStatusException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-
-    public ReadStatusException(ErrorCode errorCode, Throwable cause) {
-        super(errorCode, cause);
-    }
-}
